Add :max-queued config option to ring-jetty-adapter
When the rate of incoming requests exceeds the rate at which requests
can be processed, the request queue begins to fill. Each waiting request
holds an open socket so an unbounded queue can result in exhaustion of
file handles.

The :max-queued config option allows a limit to be placed on the size of
the incoming request queue. Once this limit is reached, new requests are
rejected immediately.